+++ control/svpwm.h
@@ -16,8 +16,8 @@
 #define DEGREE_TO_RADIAN      (3.141593/180.0)
 
 typedef struct _pwmAB_ {
-    int32_t pwm1;
-    int32_t pwm2;
+    int16_t pwm1;
+    int16_t pwm2;
 }pwmAB_t;
 
 void cal_pwmAB(pwmAB_t *pwmAB, fb_exc_angle_t *fb_exc_angle, fb_current_t *fb_current);
